requirements.txt
# ============================================================
# requirements.txt
# Install all at once: pip install -r requirements.txt
# ============================================================

# Claude API
anthropic>=0.25.0

# Web scraping (headless browser)
playwright>=1.40.0

# PDF generation
weasyprint>=60.0
markdown>=3.5.0

# Google Drive upload
google-auth>=2.25.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0

# Scheduler (for daily auto-scraping)
apscheduler>=3.10.0

# Excel reading (HR contacts + job links)
pandas>=2.0.0
openpyxl>=3.1.0

# Web scraping (JD from job URLs)
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
